requirements.txt
# requirements-gui.txt
# Core
click>=8.1.0
httpx>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.2.0
pydantic>=2.5.0
pyyaml>=6.0.1

# GUI
PyQt6>=6.5.0
qasync>=0.24.0

# Optional
browser-cookie3>=0.19.1